update utility service
