+++ src/app/components/channel-chat/channel-chat.component.html
@@ -3,15 +3,16 @@
   <section class="chat-container">
     @if(currentMessages.length == 0) {
     <div class="no-messages">{{currentChannel.name}}</div>
-    <p>Du hast diesn Channel {{formatDate(currentChannel.createdAt?.toDate())}} erstellt. Das ist der Anfang des
+    <p>Du hast diesn Channel {{utilityService.formatDate(currentChannel.createdAt?.toDate())}} erstellt. Das ist der
+      Anfang des
       Channels
       # {{currentChannel.name}}</p>
     } @else {
     @for(message of currentMessages; track message; let i = $index) {
-    @if(shouldShowTimestamp(message) || i == 0) {
+    @if(utilityService.shouldShowTimestamp(message) || i == 0) {
     <div class="date-line">
       <div class="line"></div>
-      <span class="message-date">{{ formatDate(message.sentAt.toDate()) }}</span>
+      <span class="message-date">{{ utilityService.formatDate(message.sentAt.toDate()) }}</span>
       <div class="line"></div>
     </div>
     }
